Fix kibana importer Add redis cache Drop support for unicode() Refactor StockPrice
