+++ docker-compose.yml
@@ -23,15 +23,15 @@ services:
       nproc:
         soft: 4096
         hard: 4096
-    #expose this for local dev only!
+    #expose this port for local dev only!
     ports:
       - "9200:9200"
   redis:
     build:
       context: ./redis-docker
-    #expose this for local dev only!
-    #ports:
-    #  - "6379:6379"
+    #expose this port for local dev only!
+    ports:
+      - "6379:6379"
   kibana:
     image: docker.elastic.co/kibana/kibana:7.3.1
     depends_on:
