+++ src/StockSight/Helper/Sentiment.py
@@ -7,7 +7,7 @@ from bs4 import BeautifulSoup
 from textblob import TextBlob
 from vaderSentiment.vaderSentiment import SentimentIntensityAnalyzer
 
-from Sentiment.Initializer.LoggerInit import *
+from StockSight.Initializer.Logger import *
 
 def get_page_text(url):
