+++ src/StockSight/Initializer/ElasticSearch.py
