+++ src/StockSight/Initializer/Redis.py
